# v3/services/ui/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ARM v3 Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}/transcode": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Apply Session",
        "operationId": "apply_session_api_jobs__job_id__transcode_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplySessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplySessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transcodes": {
      "get": {
        "tags": [
          "transcodes"
        ],
        "summary": "List Transcodes",
        "operationId": "list_transcodes_api_transcodes_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TranscodeTaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "session_application_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Application Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscodeTaskView"
                  },
                  "title": "Response List Transcodes Api Transcodes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transcodes/{task_id}": {
      "delete": {
        "tags": [
          "transcodes"
        ],
        "summary": "Cancel Transcode",
        "operationId": "cancel_transcode_api_transcodes__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplySessionRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "overwrite": {
            "type": "boolean",
            "title": "Overwrite",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ApplySessionRequest"
      },
      "ApplySessionResponse": {
        "properties": {
          "session_application": {
            "$ref": "#/components/schemas/SessionApplicationView"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TranscodeTaskView"
            },
            "type": "array",
            "title": "Tasks"
          },
          "collisions": {
            "items": {
              "$ref": "#/components/schemas/CollisionInfo"
            },
            "type": "array",
            "title": "Collisions"
          },
          "idempotent": {
            "type": "boolean",
            "title": "Idempotent"
          }
        },
        "type": "object",
        "required": [
          "session_application",
          "tasks",
          "collisions",
          "idempotent"
        ],
        "title": "ApplySessionResponse"
      },
...
      "CollisionInfo": {
        "properties": {
          "output_path": {
            "type": "string",
            "title": "Output Path"
          },
          "existing_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Task Id"
          },
          "on_filesystem": {
            "type": "boolean",
            "title": "On Filesystem"
          }
        },
        "type": "object",
        "required": [
          "output_path",
          "existing_task_id",
          "on_filesystem"
        ],
        "title": "CollisionInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionApplicationStatus": {
        "type": "string",
        "enum": [
          "waiting_identify",
          "queued",
          "running",
          "done",
          "done_partial",
          "failed",
          "cancelled"
        ],
        "title": "SessionApplicationStatus"
      },
      "SessionApplicationView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionApplicationStatus"
          },
          "overrides_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overrides Json"
          },
          "overwrite": {
            "type": "boolean",
            "title": "Overwrite"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "job_id",
          "status",
          "overrides_json",
          "overwrite",
          "created_by_user_id",
          "created_at",
          "completed_at"
        ],
        "title": "SessionApplicationView"
      },
...
      "TranscodeTaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "in_progress",
          "done",
          "failed"
        ],
        "title": "TranscodeTaskStatus"
      },
      "TranscodeTaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_application_id": {
            "type": "string",
            "title": "Session Application Id"
          },
          "source_track_id": {
            "type": "string",
            "title": "Source Track Id"
          },
          "status": {
            "$ref": "#/components/schemas/TranscodeTaskStatus"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "claimed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed By"
          },
          "claim_heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Heartbeat At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_application_id",
          "source_track_id",
          "status",
          "output_path",
          "progress_pct",
          "attempts",
          "claimed_by",
          "claim_heartbeat_at",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "TranscodeTaskView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}